Refactor Exercise page
